Refactor the entire PBS system...
Enable dependency handling with PBS (allows restart tests to work).
Remove PBS Emulator. The emulator does not support dependency jobs.
Refs #3152

While using PBS options, custom test parameters were not being correctly
populated in the pbs_template file. Arguments such as --timing for example
would not appear in the PBS cluster file.
Closes #8778

+++ python/ClusterLauncher/PBSJob.py
@@ -1,7 +1,7 @@
 from FactorySystem import InputParameters
 from Job import Job
 
-import os, sys, subprocess, shutil, re
+import os, sys, subprocess, shutil, re, errno
 
 class PBSJob(Job):
     def validParams():
@@ -61,17 +61,23 @@ class PBSJob(Job):
         for file in os.listdir('../'):
             if os.path.isfile('../' + file) and file != job_file and \
                (not params.isValid('no_copy') or file not in params['no_copy']) and \
-               (not params.isValid('no_copy_pattern') or pattern.match(file) is None):
+               (not params.isValid('no_copy_pattern') or pattern.match(file) is None) and \
+               not os.path.exists(file) and \
+               os.path.splitext(file)[1] != '':
                 shutil.copy('../' + file, '.')
 
         # Copy directories
         if params.isValid('copy_files'):
             for file in params['copy_files'].split():
-                print file
                 if os.path.isfile('../' + file):
-                    shutil.copy('../' + file, '.')
+                    if not os.path.exists(file):
+                        shutil.copy('../' + file, '.')
                 elif os.path.isdir('../' + file):
-                    shutil.copytree('../' + file, file)
+                    try:
+                        shutil.copytree('../' + file, file)
+                    except OSError, ex:
+                        if ex.errno == errno.EEXIST: pass
+                        else: raise
 
     def prepareJobScript(self):
         f = open(self.specs['template_script'], 'r')
